+++ packages/gov-ui-kit/src/modules/components/proposal/proposalActions/actions/proposalActionWithdrawToken/index.ts
@@ -1,0 +1,1 @@
+export { ProposalActionWithdrawToken, type IProposalActionWithdrawTokenProps } from './proposalActionWithdrawToken';