+++ packages/gov-ui-kit/src/modules/components/proposal/proposalActions/proposalActionsAction/index.ts
@@ -1,0 +1,1 @@
+export { ProposalActionsAction, type IProposalActionsActionProps } from './proposalActionsAction';